update project structure to match master

+++ ChartsDemo/Classes/RealmBase/Score.h
@@ -1,8 +1,6 @@
 //
 //  Score.h
 //  ChartsDemo
-//
-//  Created by Daniel Cohen Gindi on 17/01/2015.
 //  Copyright © 2015 dcg. All rights reserved.
 //
 
@@ -11,11 +9,11 @@
 @interface Score : RLMObject
 
 @property (nonatomic, assign) float totalScore;
-@property (nonatomic, assign) NSInteger scoreNr;
+@property (nonatomic, assign) double scoreNr;
 @property (nonatomic, strong) NSString *playerName;
 
 - (id)initWithTotalScore:(float)totalScore
-                 scoreNr:(NSInteger)scoreNr
+                 scoreNr:(double)scoreNr
               playerName:(NSString *)playerName;
 
 @end
